+++ android/src/main/java/com/example/screen_helper/ScreenSizeHelper.java
@@ -1,20 +1,13 @@
 package com.example.screen_helper;
 
-import android.view.WindowManager;
-
-import android.content.res.Configuration;
-import android.content.res.Resources;
-import android.util.DisplayMetrics;
-import android.view.Display;
-import android.view.ViewConfiguration;
-import android.view.KeyCharacterMap;
-import android.view.KeyEvent;
 import android.app.Activity;
-import android.graphics.Point;
-import android.content.Context;
-import android.view.WindowInsets;
+import android.util.DisplayMetrics;
+import android.view.WindowManager;
 import android.os.Build;
 import android.graphics.Rect;
+import android.view.Display;
+import java.util.HashMap;
+import java.util.Map;
 
 public class ScreenSizeHelper {
 
@@ -25,131 +18,40 @@ public class ScreenSizeHelper {
     public ScreenSizeHelper(Activity activity) {
         this.activity = activity;
         this.metrics = new DisplayMetrics();
+        WindowManager windowManager = activity.getWindowManager();
 
+        // Check Android version to use appropriate method for getting screen metrics
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
-            // For API level 30 and above
-            WindowManager windowManager = activity.getWindowManager();
+            // For Android 11 (API 30) and above
             Rect bounds = windowManager.getCurrentWindowMetrics().getBounds();
             metrics.widthPixels = bounds.width();
             metrics.heightPixels = bounds.height();
-            metrics.xdpi = activity.getResources().getDisplayMetrics().xdpi;
-            metrics.ydpi = activity.getResources().getDisplayMetrics().ydpi;
         } else {
-            // For API level 29 and below
-            // because of the deprecation of the getRealMetrics method
-            this.display = activity.getWindowManager().getDefaultDisplay();
+            // For Android 10 (API 29) and below
+            Display display = windowManager.getDefaultDisplay();
             display.getRealMetrics(metrics);
         }
+
+        // Set the screen density (DPI) values
+        metrics.xdpi = activity.getResources().getDisplayMetrics().xdpi;
+        metrics.ydpi = activity.getResources().getDisplayMetrics().ydpi;
     }
 
-    public double getScreenPPI() {
-        // Get the diagonal size in pixels
-        double diagonalPixels = getScreenDiagonalInPixels();
-
-        // Get the diagonal size in inches
-        double diagonalInches = getScreenDiagonalInInches();
-
-        // Calculate the PPI
-        return diagonalPixels / diagonalInches;
-    }
-
-    public double getScreenWidthInInches() {
-
-        // Get the screen width in pixels
-        int widthPixels = getScreenRealWidthInPixels();
-
-        // Calculate the width in inches
-        return widthPixels / metrics.xdpi;
-    }
-
-    public double getScreenHeightInInches() {
-
-        // Get the screen height in pixels
-        int heightPixels = getScreenRealHeightInPixels();
-
-        // Calculate the height in inches
-        return heightPixels / metrics.ydpi;
-    }
-
-    // Get the screen width in pixels
-    public int getScreenRealWidthInPixels() {
-        Resources resources = activity.getResources();
-
-        return metrics.widthPixels;
-
-    }
-
-    // Get the screen height in pixels
-    public int getScreenRealHeightInPixels() {
-        Resources resources = activity.getResources();
-
-        return metrics.heightPixels;
-    }
-
-    public int getScreenDiagonalInPixels() {
-
-        int widthPixels = getScreenRealWidthInPixels();
-        int heightPixels = getScreenRealHeightInPixels();
-
-        return (int) Math.sqrt(Math.pow(widthPixels, 2) + Math.pow(heightPixels, 2));
-    }
-
-    public double getScreenDiagonalInInches() {
-
-        // Get the screen dimensions in pixels
-        int widthPixels = getScreenRealWidthInPixels();
-        int heightPixels = getScreenRealHeightInPixels();
-
-        // Get the diagonal size in pixels
-        double diagonalPixels = getScreenDiagonalInPixels();
-
-        // Get the diagonal size in inches
-        double widthInches = getScreenWidthInInches();
-        double heightInches = getScreenHeightInInches();
-        return Math.sqrt(Math.pow(widthInches, 2) + Math.pow(heightInches, 2));
-    }
-
-    public int getStatusBarHeight() {
-        int result = 0;
-        int resourceId = activity.getResources().getIdentifier("status_bar_height", "dimen", "android");
-        if (resourceId > 0) {
-            result = activity.getResources().getDimensionPixelSize(resourceId);
-        }
+    // function to get real screen size in inches (width and height)
+    public Map<String, Double> getScreenSizeInInches() {
+        double widthInches = metrics.widthPixels / metrics.xdpi;
+        double heightInches = metrics.heightPixels / metrics.ydpi;
+        Map<String, Double> result = new HashMap<>();
+        result.put("width", widthInches);
+        result.put("height", heightInches);
         return result;
     }
 
-    public int getNavigationBarHeight11() {
-        if (Build.VERSION.SDK_INT >= 30) {
-            WindowInsets windowInsets = activity.getWindowManager().getCurrentWindowMetrics().getWindowInsets();
-            return windowInsets.getInsets(WindowInsets.Type.navigationBars()).bottom;
-        } else {
-            return calculateNavigationBarHeightForOlderDevices();
-        }
+    // function to get screen resolution in pixels (width and height)
+    public Map<String, Double> screenSizeInPixels() {
+        Map<String, Double> result = new HashMap<>();
+        result.put("width", Double.valueOf(metrics.widthPixels));
+        result.put("height", Double.valueOf(metrics.heightPixels));
+        return result;
     }
-
-    private int calculateNavigationBarHeightForOlderDevices() {
-        Display display = activity.getWindowManager().getDefaultDisplay();
-        Point appUsableSize = new Point();
-        Point realScreenSize = new Point();
-        display.getSize(appUsableSize);
-        display.getRealSize(realScreenSize);
-
-        // navigation bar on the side
-        if (appUsableSize.x < realScreenSize.x) {
-            return realScreenSize.x - appUsableSize.x;
-        }
-
-        // navigation bar at the bottom
-        if (appUsableSize.y < realScreenSize.y) {
-            return realScreenSize.y - appUsableSize.y;
-        }
-
-        return 0;
-    }
-
-    private boolean isTablet() {
-        return (activity.getResources().getConfiguration().screenLayout
-                & Configuration.SCREENLAYOUT_SIZE_MASK) >= Configuration.SCREENLAYOUT_SIZE_LARGE;
-    }
-
 }