+++ lib/screen_helper_method_channel.dart
@@ -9,73 +9,31 @@ class MethodChannelScreenHelper extends ScreenHelperPlatform {
   @visibleForTesting
   final methodChannel = const MethodChannel('screen_helper');
 
-  /// Fetches the screen's PPI (Pixels Per Inch).
+  /// Fetches the screen size in inches (width and height).
   @override
-  Future<double?> getScreenPPI() async {
-    final screenPPI = await methodChannel.invokeMethod<double>('getScreenPPI');
-
-    return screenPPI;
-  }
-
-  /// Fetches the screen's diagonal size in inches.
-  @override
-  Future<double?> getScreenDiagonalInInches() async {
-    final screenDiagonalInInches = await methodChannel
-        .invokeMethod<double>('getScreenDiagonalSizeInInches');
-
-    return screenDiagonalInInches;
-  }
-
-  /// Fetches the screen's width size in inches.
-  @override
-  Future<double?> getScreenWidthInInches() async {
-    final screenWidthInInches =
-        await methodChannel.invokeMethod<double>('getScreenWidthSizeInInches');
-
-    if (screenWidthInInches != null) {
-      return double.parse(screenWidthInInches.toStringAsFixed(3));
+  Future<Map<String, double>?> getScreenSizeInInches() async {
+    final Map<dynamic, dynamic>? result = await methodChannel
+        .invokeMethod<Map<dynamic, dynamic>>('getScreenSizeInInches');
+    if (result != null) {
+      return {
+        'width': result['width'] as double,
+        'height': result['height'] as double,
+      };
     }
-
     return null;
   }
 
-  /// Fetches the screen's height size in inches.
+  /// Fetches the screen resolution in pixels (width and height).
   @override
-  Future<double?> getScreenHeightInInches() async {
-    final screenHeightInInches =
-        await methodChannel.invokeMethod<double>('getScreenHeightSizeInInches');
-
-    if (screenHeightInInches != null) {
-      return double.parse(screenHeightInInches.toStringAsFixed(3));
+  Future<Map<String, double>?> getScreenResolution() async {
+    final Map<dynamic, dynamic>? result = await methodChannel
+        .invokeMethod<Map<dynamic, dynamic>>('getScreenResolution');
+    if (result != null) {
+      return {
+        'width': result['width'],
+        'height': result['height'],
+      };
     }
-
     return null;
-  }
-
-  /// Fetches the screen's real width in pixels.
-  @override
-  Future<int?> getScreenRealWidthInPixels() async {
-    final screenRealWidthInPixels =
-        await methodChannel.invokeMethod<int>('getScreenRealWidthInPixels');
-
-    return screenRealWidthInPixels;
-  }
-
-  /// Fetches the screen's real height in pixels.
-  @override
-  Future<int?> getScreenRealHeightInPixels() async {
-    final screenRealHeightInPixels =
-        await methodChannel.invokeMethod<int>('getScreenRealHeightInPixels');
-
-    return screenRealHeightInPixels;
-  }
-
-  /// Fetches the screen's diagonal size in pixels.
-  @override
-  Future<int?> getScreenDiagonalInPixels() async {
-    final screenDiagonalInPixels =
-        await methodChannel.invokeMethod<int?>('screenDiagonalSizeInPixels');
-
-    return screenDiagonalInPixels;
   }
 }