+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 1.0.0
+* changed the implementation to use inherited widget. so you can access the screen information from any widget in the widget tree.
+* changed method to convert to be an extension method. so you can use it like `context.mmToPx(15)`
+* added support to get the screen diagonal in pixels.
+* added support to convert pixels to mm.
+* added support to convert inches to pixels and pixels to inches.
+* added support to convert cm to pixels and pixels to cm.
+
+
+
+
 ## 0.0.3
 
 * Updated README.md file.